+++ src/templates/layouts/index.ts
@@ -1,3 +1,4 @@
 export * from './layout-template.catalog'
+export * from './main-layout-persistence'
 export { default as AuthLayoutTemplate } from './AuthLayoutTemplate.vue'
 export { default as MainLayoutTemplate } from './MainLayoutTemplate.vue'